Re-extract locals to common
